Initial dump of many components for the game
This includes the board and the lesser structs within, checking of the
game-ending conditions, as well as movement logic and ptn parsing (though
these are less thoroughly tested). The rules of the game are kept in a
trait called RuleSet in order to keep the rules generic enough to allow
variants to be constructed with relative ease. Most of the essentials for
Tak are present now, but they are disjointed and work will have to be
done to integrate them.